Added mounting process.
