--- app/javascript/packs/vm-snapshot-form.js
@@ -1,1 +1,0 @@
-import '../vm-snapshot-form/register';
